Updated api and dashboard

+++ sadet.autodb.api/Interfaces/IGamesRepository.cs
@@ -4,7 +4,7 @@ namespace sadet.autodb.api.Interfaces;
 
 public interface IGamesRepository
 {
-    public bool Exists(int appid);
+    public bool Exists(long appid);
     public ICollection<Game> Get();
-    public Game Get(int appid);
+    public Game Get(long appid);
 }
